Add TinyintType.getByte to directly fetch value as a byte

+++ core/trino-main/src/main/java/io/trino/operator/MergeWriterOperator.java
@@ -152,7 +152,7 @@ public class MergeWriterOperator
         long insertsFromUpdates = 0;
         int positionCount = page.getPositionCount();
         for (int position = 0; position < positionCount; position++) {
-            insertsFromUpdates += TINYINT.getLong(insertFromUpdateColumn, position);
+            insertsFromUpdates += TINYINT.getByte(insertFromUpdateColumn, position);
         }
         rowCount += positionCount - insertsFromUpdates;
     }
